+++ devtools/travis-ci/install.sh
@@ -17,7 +17,7 @@ bash $MINICONDA -b -p $MINICONDA_HOME
 export PIP_ARGS="-U"
 export PATH=$MINICONDA_HOME/bin:$PATH
 conda update --yes conda
-conda install --yes conda-build=2.1.5 jinja2 anaconda-client pip 
+conda install --yes conda-build jinja2 anaconda-client pip 
 conda install --yes -c omnia openmoltools
 conda install --yes -c omnia parmed
 conda install --yes -c matplotlib
