+++ openforcefield/typing/__init__.py
@@ -1,0 +1,1 @@
+# Typing tools for forcefields
