+++ openforcefield/typing/chemistry/__init__.py
@@ -1,0 +1,3 @@
+# Tools dealing with chemical environments or chemistry
+
+from openforcefield.typing.chemistry.environment import *
