+++ openforcefield/typing/engines/__init__.py
@@ -1,0 +1,1 @@
+# Engines for forcefield typing
