+++ openforcefield/utils/__init__.py
@@ -1,0 +1,3 @@
+# General utilities for forcefields
+
+from openforcefield.utils.utils import *
